rp vs porb: P22 params only; vE18 line maybe off
Vincent Van Eylen's asteroseismic line is for pretty old stars.  You
might expect it to fall at a different location for younger stars (eg.
as David+2021 argued).

Also, the issue of whether you use the Petigura+2022 parameters alone,
versus the Berger+2020 parameters *in combination* with the
Petigura+2022 parameters, makes a significant difference on the clarify
of the valley.

For the Petigura+2022 parameters alone, there are like 1.5-2x fewer
planets, but the valley looks cleaner.